requirements.txt
spotipy>=2.25.1
pillow>=11.2.1
streamdeck>=0.9.6
python-dotenv>=1.1.0
